+++ webapi/src/main/resources/application.conf
@@ -165,23 +165,9 @@ akka.actor.deployment {
         }
     }
 
-    /responderManager/sipiRouterV1 {
+    /storeManager/sipiManager/sipiConnector {
         router = balancing-pool
-        nr-of-instances = 5
-        pool-dispatcher {
-            executor = "thread-pool-executor"
-
-            # allocate exactly 5 threads for this pool
-            thread-pool-executor {
-                core-pool-size-min = 5
-                core-pool-size-max = 5
-            }
-        }
-    }
-
-    /responderManager/sipiRouterV2 {
-        router = balancing-pool
-        nr-of-instances = 5
+        nr-of-instances = 10
         pool-dispatcher {
             executor = "thread-pool-executor"
