+++ webapi/src/main/scala/org/knora/webapi/routing/v2/ResourcesRouteV2.scala
@@ -30,7 +30,7 @@ import org.knora.webapi._
 import org.knora.webapi.messages.v2.responder.resourcemessages._
 import org.knora.webapi.responders.RESPONDER_MANAGER_ACTOR_PATH
 import org.knora.webapi.routing.{Authenticator, RouteUtilV2}
-import org.knora.webapi.store.STORE_MANAGER_ACTOR_PATH
+import org.knora.webapi.store.StoreManagerActorPath
 import org.knora.webapi.util.IriConversions._
 import org.knora.webapi.util.jsonld.{JsonLDDocument, JsonLDUtil}
 import org.knora.webapi.util.{SmartIri, StringFormatter}
@@ -137,7 +137,7 @@ object ResourcesRouteV2 extends Authenticator {
         implicit val timeout: Timeout = settings.defaultTimeout
         implicit val stringFormatter: StringFormatter = StringFormatter.getGeneralInstance
         val responderManager = system.actorSelection(RESPONDER_MANAGER_ACTOR_PATH)
-        val storeManager = system.actorSelection(STORE_MANAGER_ACTOR_PATH)
+        val storeManager = system.actorSelection(StoreManagerActorPath)
 
         path("v2" / "resources") {
             post {
