+++ webapi/src/main/scala/org/knora/webapi/routing/v2/StandoffRouteV2.scala
@@ -32,7 +32,7 @@ import akka.util.Timeout
 import org.knora.webapi.messages.v2.responder.standoffmessages.{CreateMappingRequestMetadataV2, CreateMappingRequestV2, CreateMappingRequestXMLV2}
 import org.knora.webapi.responders.RESPONDER_MANAGER_ACTOR_PATH
 import org.knora.webapi.routing.{Authenticator, RouteUtilV2}
-import org.knora.webapi.store.STORE_MANAGER_ACTOR_PATH
+import org.knora.webapi.store.StoreManagerActorPath
 import org.knora.webapi.util.StringFormatter
 import org.knora.webapi.util.jsonld.JsonLDUtil
 import org.knora.webapi.{ApiV2WithValueObjects, BadRequestException, KnoraDispatchers, SettingsImpl}
@@ -52,7 +52,7 @@ object StandoffRouteV2 extends Authenticator {
         implicit val stringFormatter: StringFormatter = StringFormatter.getGeneralInstance
         implicit val materializer: ActorMaterializer = ActorMaterializer()
         val responderManager = system.actorSelection(RESPONDER_MANAGER_ACTOR_PATH)
-        val storeManager = system.actorSelection(STORE_MANAGER_ACTOR_PATH)
+        val storeManager = system.actorSelection(StoreManagerActorPath)
 
         path("v2" / "mapping") {
             post {
