+++ webapi/src/main/scala/org/knora/webapi/routing/v2/ValuesRouteV2.scala
@@ -30,7 +30,7 @@ import org.knora.webapi._
 import org.knora.webapi.messages.v2.responder.valuemessages._
 import org.knora.webapi.responders.RESPONDER_MANAGER_ACTOR_PATH
 import org.knora.webapi.routing.{Authenticator, RouteUtilV2}
-import org.knora.webapi.store.STORE_MANAGER_ACTOR_PATH
+import org.knora.webapi.store.StoreManagerActorPath
 import org.knora.webapi.util.IriConversions._
 import org.knora.webapi.util.StringFormatter
 import org.knora.webapi.util.jsonld.{JsonLDDocument, JsonLDUtil}
@@ -47,7 +47,7 @@ object ValuesRouteV2 extends Authenticator {
         implicit val timeout: Timeout = settings.defaultTimeout
         implicit val stringFormatter: StringFormatter = StringFormatter.getGeneralInstance
         val responderManager = system.actorSelection(RESPONDER_MANAGER_ACTOR_PATH)
-        val storeManager = system.actorSelection(STORE_MANAGER_ACTOR_PATH)
+        val storeManager = system.actorSelection(StoreManagerActorPath)
 
         // #post-value-parse-jsonld
         path("v2" / "values") {
