+++ webapi/src/main/scala/org/knora/webapi/store/sipi/SipiConnector.scala
@@ -21,32 +21,39 @@ package org.knora.webapi.store.sipi
 
 import java.util
 
-import akka.actor.{Actor, ActorLogging}
-import akka.stream.ActorMaterializer
+import akka.actor.{Actor, ActorLogging, ActorSystem}
 import org.apache.http.client.config.RequestConfig
 import org.apache.http.client.entity.UrlEncodedFormEntity
-import org.apache.http.client.methods.HttpPost
+import org.apache.http.client.methods.{CloseableHttpResponse, HttpDelete, HttpGet, HttpPost}
 import org.apache.http.client.protocol.HttpClientContext
 import org.apache.http.impl.client.{CloseableHttpClient, HttpClients}
 import org.apache.http.message.BasicNameValuePair
 import org.apache.http.util.EntityUtils
-import org.apache.http.{HttpHost, HttpRequest, NameValuePair}
-import org.knora.webapi.messages.store.sipimessages.{DeleteTemporaryFileRequestV2, GetImageMetadataRequestV2, MoveTemporaryFileToPermanentStorageRequestV2}
+import org.apache.http.{Consts, HttpHost, HttpRequest, NameValuePair}
+import org.knora.webapi.messages.store.sipimessages.SipiConstants.FileType
+import org.knora.webapi.messages.store.sipimessages._
+import org.knora.webapi.messages.v1.responder.valuemessages.{FileValueV1, StillImageFileValueV1, TextFileValueV1}
 import org.knora.webapi.messages.v2.responder.SuccessResponseV2
-import org.knora.webapi.responders.Responder
 import org.knora.webapi.routing.JWTHelper
 import org.knora.webapi.util.ActorUtil.{handleUnexpectedMessage, try2Message}
-import org.knora.webapi.util.SipiUtil
-import org.knora.webapi.{BadRequestException, SipiException}
+import org.knora.webapi.util.{SipiUtil, StringFormatter}
+import org.knora.webapi.{BadRequestException, KnoraDispatchers, NotImplementedException, Settings, SipiException}
 import spray.json._
 
+import scala.concurrent.ExecutionContext
 import scala.util.Try
 
 /**
   * Makes requests to Sipi.
   */
 class SipiConnector extends Actor with ActorLogging {
-    implicit val materializer: ActorMaterializer = ActorMaterializer()
+
+    implicit val system: ActorSystem = context.system
+    implicit val executionContext: ExecutionContext = system.dispatchers.lookup(KnoraDispatchers.KnoraActorDispatcher)
+
+    val settings = Settings(system)
+
+    implicit val stringFormatter: StringFormatter = StringFormatter.getGeneralInstance
 
     private val targetHost: HttpHost = new HttpHost(settings.internalSipiHost, settings.internalSipiPort, "http")
 
@@ -61,10 +68,174 @@ class SipiConnector extends Actor with ActorLogging {
     private val httpClient: CloseableHttpClient = HttpClients.custom.setDefaultRequestConfig(sipiRequestConfig).build
 
     override def receive: Receive = {
+        case convertPathRequest: SipiResponderConversionPathRequestV1 => try2Message(sender(), convertPathV1(convertPathRequest), log)
+        case convertFileRequest: SipiResponderConversionFileRequestV1 => try2Message(sender(), convertFileV1(convertFileRequest), log)
         case getFileMetadataRequestV2: GetImageMetadataRequestV2 => try2Message(sender(), getFileMetadataV2(getFileMetadataRequestV2), log)
         case moveTemporaryFileToPermanentStorageRequestV2: MoveTemporaryFileToPermanentStorageRequestV2 => try2Message(sender(), moveTemporaryFileToPermanentStorageV2(moveTemporaryFileToPermanentStorageRequestV2), log)
         case deleteTemporaryFileRequestV2: DeleteTemporaryFileRequestV2 => try2Message(sender(), deleteTemporaryFileV2(deleteTemporaryFileRequestV2), log)
         case other => handleUnexpectedMessage(sender(), other, log, this.getClass.getName)
+    }
+
+    /**
+      * Convert a file that has been sent to Knora (non GUI-case).
+      *
+      * @param conversionRequest the information about the file (uploaded by Knora).
+      * @return a [[SipiResponderConversionResponseV1]] representing the file values to be added to the triplestore.
+      */
+    private def convertPathV1(conversionRequest: SipiResponderConversionPathRequestV1): Try[SipiResponderConversionResponseV1] = {
+        val url = s"${settings.internalSipiImageConversionUrlV1}/${settings.sipiPathConversionRouteV1}"
+
+        callSipiConvertRoute(url, conversionRequest)
+    }
+
+    /**
+      * Convert a file that is already managed by Sipi (GUI-case).
+      *
+      * @param conversionRequest the information about the file (managed by Sipi).
+      * @return a [[SipiResponderConversionResponseV1]] representing the file values to be added to the triplestore.
+      */
+    private def convertFileV1(conversionRequest: SipiResponderConversionFileRequestV1): Try[SipiResponderConversionResponseV1] = {
+        val url = s"${settings.internalSipiImageConversionUrlV1}/${settings.sipiFileConversionRouteV1}"
+
+        callSipiConvertRoute(url, conversionRequest)
+    }
+
+    /**
+      * Makes a conversion request to Sipi and creates a [[SipiResponderConversionResponseV1]]
+      * containing the file values to be added to the triplestore.
+      *
+      * @param urlPath           the Sipi route to be called.
+      * @param conversionRequest the message holding the information to make the request.
+      * @return a [[SipiResponderConversionResponseV1]].
+      */
+    private def callSipiConvertRoute(urlPath: String, conversionRequest: SipiResponderConversionRequestV1): Try[SipiResponderConversionResponseV1] = {
+        val context: HttpClientContext = HttpClientContext.create
+
+        val formParams = new util.ArrayList[NameValuePair]()
+
+        for ((key, value) <- conversionRequest.toFormData()) {
+            formParams.add(new BasicNameValuePair(key, value))
+        }
+
+        val postEntity = new UrlEncodedFormEntity(formParams, Consts.UTF_8)
+        val httpPost = new HttpPost(urlPath)
+        httpPost.setEntity(postEntity)
+
+        val conversionResultTry: Try[String] = Try {
+            var maybeResponse: Option[CloseableHttpResponse] = None
+
+            try {
+                maybeResponse = Some(httpClient.execute(targetHost, httpPost, context))
+
+                val responseEntityStr: String = Option(maybeResponse.get.getEntity) match {
+                    case Some(responseEntity) => EntityUtils.toString(responseEntity)
+                    case None => ""
+                }
+
+                val statusCode: Int = maybeResponse.get.getStatusLine.getStatusCode
+                val statusCategory: Int = statusCode / 100
+
+                // Was the request successful?
+                if (statusCategory == 2) {
+                    // Yes.
+                    responseEntityStr
+                } else {
+                    // No. Throw an appropriate exception.
+                    val sipiErrorMsg = SipiUtil.getSipiErrorMessage(responseEntityStr)
+
+                    if (statusCategory == 4) {
+                        throw BadRequestException(s"Sipi responded with HTTP status code $statusCode: $sipiErrorMsg")
+                    } else {
+                        throw SipiException(s"Sipi responded with HTTP status code $statusCode: $sipiErrorMsg")
+                    }
+                }
+            } finally {
+                maybeResponse match {
+                    case Some(response) => response.close()
+                    case None => ()
+                }
+            }
+        }
+
+        //
+        // handle unsuccessful requests to Sipi
+        //
+        val recoveredConversionResultTry = conversionResultTry.recoverWith {
+            case badRequestException: BadRequestException => throw badRequestException
+            case sipiException: SipiException => throw sipiException
+            case e: Exception => throw SipiException("Failed to connect to Sipi", e, log)
+        }
+
+        for {
+            responseAsStr: String <- recoveredConversionResultTry
+
+            /* get json from response body */
+            responseAsJson: JsValue = JsonParser(responseAsStr)
+
+            // get file type from Sipi response
+            fileType: String = responseAsJson.asJsObject.fields.getOrElse("file_type", throw SipiException(message = "Sipi did not return a file type")) match {
+                case JsString(ftype: String) => ftype
+                case other => throw SipiException(message = s"Sipi returned an invalid file type: $other")
+            }
+
+            // turn fileType returned by Sipi (a string) into an enum
+            fileTypeEnum: FileType.Value = SipiConstants.FileType.lookup(fileType)
+
+            // create the apt case class depending on the file type returned by Sipi
+            fileValuesV1: Vector[FileValueV1] = fileTypeEnum match {
+                case SipiConstants.FileType.IMAGE =>
+                    // parse response as a [[SipiImageConversionResponse]]
+                    val imageConversionResult = try {
+                        responseAsJson.convertTo[SipiImageConversionResponse]
+                    } catch {
+                        case e: DeserializationException => throw SipiException(message = "JSON response returned by Sipi is invalid, it cannot be turned into a SipiImageConversionResponse", e = e, log = log)
+                    }
+
+                    // create two StillImageFileValueV1s
+                    Vector(StillImageFileValueV1(// full representation
+                        internalMimeType = stringFormatter.toSparqlEncodedString(imageConversionResult.mimetype_full, throw BadRequestException(s"The internal MIME type returned by Sipi is invalid: '${imageConversionResult.mimetype_full}")),
+                        originalFilename = stringFormatter.toSparqlEncodedString(imageConversionResult.original_filename, throw BadRequestException(s"The original filename returned by Sipi is invalid: '${imageConversionResult.original_filename}")),
+                        originalMimeType = Some(stringFormatter.toSparqlEncodedString(imageConversionResult.original_mimetype, throw BadRequestException(s"The original MIME type returned by Sipi is invalid: '${imageConversionResult.original_mimetype}"))),
+                        dimX = imageConversionResult.nx_full,
+                        dimY = imageConversionResult.ny_full,
+                        internalFilename = stringFormatter.toSparqlEncodedString(imageConversionResult.filename_full, throw BadRequestException(s"The internal filename returned by Sipi is invalid: '${imageConversionResult.filename_full}")),
+                        qualityLevel = 100,
+                        qualityName = Some(SipiConstants.StillImage.fullQuality)
+                    ),
+                        StillImageFileValueV1(// thumbnail representation
+                            internalMimeType = stringFormatter.toSparqlEncodedString(imageConversionResult.mimetype_thumb, throw BadRequestException(s"The internal MIME type returned by Sipi is invalid: '${imageConversionResult.mimetype_full}")),
+                            originalFilename = stringFormatter.toSparqlEncodedString(imageConversionResult.original_filename, throw BadRequestException(s"The original filename returned by Sipi is invalid: '${imageConversionResult.original_filename}")),
+                            originalMimeType = Some(stringFormatter.toSparqlEncodedString(imageConversionResult.original_mimetype, throw BadRequestException(s"The original MIME type returned by Sipi is invalid: '${imageConversionResult.original_mimetype}"))),
+                            dimX = imageConversionResult.nx_thumb,
+                            dimY = imageConversionResult.ny_thumb,
+                            internalFilename = stringFormatter.toSparqlEncodedString(imageConversionResult.filename_thumb, throw BadRequestException(s"The internal filename returned by Sipi is invalid: '${imageConversionResult.filename_thumb}")),
+                            qualityLevel = 10,
+                            qualityName = Some(SipiConstants.StillImage.thumbnailQuality),
+                            isPreview = true
+                        ))
+
+                case SipiConstants.FileType.TEXT =>
+
+                    // parse response as a [[SipiTextResponse]]
+                    val textStoreResult = try {
+                        responseAsJson.convertTo[SipiTextResponse]
+                    } catch {
+                        case e: DeserializationException => throw SipiException(message = "JSON response returned by Sipi is invalid, it cannot be turned into a SipiTextResponse", e = e, log = log)
+                    }
+
+                    Vector(TextFileValueV1(
+                        internalMimeType = stringFormatter.toSparqlEncodedString(textStoreResult.mimetype, throw BadRequestException(s"The internal MIME type returned by Sipi is invalid: '${textStoreResult.mimetype}")),
+                        internalFilename = stringFormatter.toSparqlEncodedString(textStoreResult.filename, throw BadRequestException(s"The internal filename returned by Sipi is invalid: '${textStoreResult.filename}")),
+                        originalFilename = stringFormatter.toSparqlEncodedString(textStoreResult.original_filename, throw BadRequestException(s"The internal filename returned by Sipi is invalid: '${textStoreResult.original_filename}")),
+                        originalMimeType = Some(stringFormatter.toSparqlEncodedString(textStoreResult.mimetype, throw BadRequestException(s"The orignal MIME type returned by Sipi is invalid: '${textStoreResult.original_mimetype}")))
+                    ))
+
+                case unknownType => throw NotImplementedException(s"Could not handle file type $unknownType")
+
+                // TODO: add missing file types
+            }
+
+        } yield SipiResponderConversionResponseV1(fileValuesV1, file_type = fileTypeEnum)
     }
 
     /**
