+++ webapi/src/test/scala/org/knora/webapi/CoreSpec.scala
@@ -29,7 +29,7 @@ import org.knora.webapi.app.{APPLICATION_STATE_ACTOR_NAME, ApplicationStateActor
 import org.knora.webapi.messages.store.triplestoremessages.{RdfDataObject, ResetTriplestoreContent}
 import org.knora.webapi.messages.v1.responder.ontologymessages.LoadOntologiesRequest
 import org.knora.webapi.responders.{MockableResponderManager, RESPONDER_MANAGER_ACTOR_NAME}
-import org.knora.webapi.store.{STORE_MANAGER_ACTOR_NAME, StoreManager}
+import org.knora.webapi.store.{StoreManagerActorName, StoreManager}
 import org.knora.webapi.util.{CacheUtil, StringFormatter}
 import org.scalatest.{BeforeAndAfterAll, Matchers, WordSpecLike}
 
@@ -75,7 +75,7 @@ abstract class CoreSpec(_system: ActorSystem) extends TestKit(_system) with Word
     lazy val mockResponders: Map[String, ActorRef] = Map.empty[String, ActorRef]
 
     val applicationStateActor: ActorRef = system.actorOf(Props(new ApplicationStateActor), name = APPLICATION_STATE_ACTOR_NAME)
-    val storeManager: ActorRef = system.actorOf(Props(new StoreManager with LiveActorMaker), name = STORE_MANAGER_ACTOR_NAME)
+    val storeManager: ActorRef = system.actorOf(Props(new StoreManager with LiveActorMaker), name = StoreManagerActorName)
     val responderManager: ActorRef = system.actorOf(Props(new MockableResponderManager(mockResponders, applicationStateActor, storeManager)), name = RESPONDER_MANAGER_ACTOR_NAME)
